+++ src/Ocelot/Cache/Middleware/OutputCacheMiddleware.cs
@@ -14,7 +14,7 @@ namespace Ocelot.Cache.Middleware
             IOcelotLoggerFactory loggerFactory,
             IOcelotCache<CachedResponse> outputCache,
             ICacheKeyGenerator cacheGenerator)
-                : base(loggerFactory.CreateLogger<OutputCacheMiddleware>())
+            : base(loggerFactory.CreateLogger<OutputCacheMiddleware>())
         {
             _next = next;
             _outputCache = outputCache;
@@ -36,18 +36,13 @@ namespace Ocelot.Cache.Middleware
             var downStreamRequestCacheKey = await _cacheGenerator.GenerateRequestCacheKey(downstreamRequest, downstreamRoute);
 
             Logger.LogDebug(() => $"Started checking cache for the '{downstreamUrlKey}' key.");
-
             var cached = _outputCache.Get(downStreamRequestCacheKey, downstreamRoute.CacheOptions.Region);
-
             if (cached != null)
             {
                 Logger.LogDebug(() => $"Cache entry exists for the '{downstreamUrlKey}' key.");
-
                 var response = CreateHttpResponseMessage(cached);
                 SetHttpResponseMessageThisRequest(httpContext, response);
-
                 Logger.LogDebug(() => $"Finished returning of cached response for the '{downstreamUrlKey}' key.");
-
                 return;
             }
 
@@ -58,24 +53,18 @@ namespace Ocelot.Cache.Middleware
             if (httpContext.Items.Errors().Count > 0)
             {
                 Logger.LogDebug(() => $"There was a pipeline error for the '{downstreamUrlKey}' key.");
-
                 return;
             }
 
             var downstreamResponse = httpContext.Items.DownstreamResponse();
-
             cached = await CreateCachedResponse(downstreamResponse);
 
             _outputCache.Add(downStreamRequestCacheKey, cached, TimeSpan.FromSeconds(downstreamRoute.CacheOptions.TtlSeconds), downstreamRoute.CacheOptions.Region);
-
             Logger.LogDebug(() => $"Finished response added to cache for the '{downstreamUrlKey}' key.");
         }
 
-        private static void SetHttpResponseMessageThisRequest(HttpContext context,
-                                                       DownstreamResponse response)
-        {
-            context.Items.UpsertDownstreamResponse(response);
-        }
+        private static void SetHttpResponseMessageThisRequest(HttpContext context, DownstreamResponse response)
+            => context.Items.UpsertDownstreamResponse(response);
 
         internal DownstreamResponse CreateHttpResponseMessage(CachedResponse cached)
         {
@@ -85,7 +74,6 @@ namespace Ocelot.Cache.Middleware
             }
 
             var content = new MemoryStream(Convert.FromBase64String(cached.Body));
-
             var streamContent = new StreamContent(content);
 
             foreach (var header in cached.ContentHeaders)
@@ -114,7 +102,6 @@ namespace Ocelot.Cache.Middleware
             }
 
             var contentHeaders = response?.Content?.Headers.ToDictionary(v => v.Key, v => v.Value);
-
             var cached = new CachedResponse(statusCode, headers, body, contentHeaders, response.ReasonPhrase);
             return cached;
         }
